+++ plom/client/annotator.py
@@ -228,6 +228,21 @@ class Annotator(QWidget):
         m.addAction("Compact UI\thome", self.narrowLayout)
         m.addAction("&Wide UI\thome", self.wideLayout)
         m.addSeparator()
+        m.addAction(
+            "Increase annotation scale\tshift-]", lambda: self.change_annot_scale(1.1)
+        )
+        # Keep a reference to this one so we can update the text
+        self._reset_scale_menu_text = "Reset annotation scale"
+        self._reset_scale_QAction = m.addAction(
+            self._reset_scale_menu_text, self.change_annot_scale
+        )
+        self.update_annot_scale_menu_label()
+
+        m.addAction(
+            "Decrease annotation scale\tshift-]",
+            lambda: self.change_annot_scale(1.0 / 1.1),
+        )
+        m.addSeparator()
         m.addAction("Help", self.menuDummy).setEnabled(False)
         m.addAction("Show shortcut keys...\t?", self.keyPopUp)
         m.addAction("About Plom", self.menuDummy).setEnabled(False)
@@ -418,6 +433,35 @@ class Annotator(QWidget):
 
         # reset the timer (its not needed to make a new one)
         self.timer.start()
+
+    def change_annot_scale(self, scale=None):
+        """Change the scale of the annotations.
+
+        args:
+            scale (float/None): if None reset the scale to the default.
+                If any floating point number, multiple the scale by that
+                value.
+        """
+        if scale is None:
+            log.info("resetting annotation scale to default")
+            if self.scene:
+                self.scene.reset_scale_factor()
+            self.update_annot_scale_menu_label()
+
+            return
+        log.info("multiplying annotation scale by {}".format(scale))
+        if self.scene:
+            self.scene.increase_scale_factor(scale)
+        self.update_annot_scale_menu_label()
+
+    def update_annot_scale_menu_label(self):
+        """Update the menu which shows the current annotation scale."""
+        if not self.scene:
+            return
+        self._reset_scale_QAction.setText(
+            self._reset_scale_menu_text
+            + "\t{:.0%}".format(self.scene.get_scale_factor())
+        )
 
     def setCurrentMarkMode(self):
         """
@@ -1068,6 +1112,15 @@ class Annotator(QWidget):
         self.zoomToggleShortCut = QShortcut(QKeySequence("Ctrl+="), self)
         self.zoomToggleShortCut.activated.connect(self.view.zoomToggle)
 
+        self.scaleAnnotIncShortCut = QShortcut(QKeySequence("Shift+]"), self)
+        self.scaleAnnotIncShortCut.activated.connect(
+            lambda: self.change_annot_scale(1.1)
+        )
+        self.scaleAnnotDecShortCut = QShortcut(QKeySequence("Shift+["), self)
+        self.scaleAnnotDecShortCut.activated.connect(
+            lambda: self.change_annot_scale(1 / 1.1)
+        )
+
         # shortcuts for undo/redo
         self.undoShortCut = QShortcut(QKeySequence("Ctrl+z"), self)
         self.undoShortCut.activated.connect(self.undo)
@@ -1691,6 +1744,7 @@ class Annotator(QWidget):
             "markStyle": self.markStyle,
             "maxMark": self.maxMark,
             "currentMark": self.score,
+            "sceneScale": self.scene.get_scale_factor(),
             "sceneItems": lst,
         }
         # save pickled file as <blah>.plom
@@ -1712,6 +1766,8 @@ class Annotator(QWidget):
 
         """
         self.view.setHidden(True)
+        if plomData.get("sceneScale", None):
+            self.scene.set_scale_factor(plomData["sceneScale"])
         self.scene.unpickleSceneItems(plomData["sceneItems"])
         # if markstyle is "Total", then click appropriate button
         if self.markStyle == 1: